Add fully parametric plate generation
Currently, the plates are generated by importing one of several STL
files to use as the base geometry for the plate. The selection of
which STL to import depends upon the length of the serial number and
whether or not a logo is desired.

Rather than relying on the pre-generated base STLs, the full geometry
of the plate can be generated parametrically. This change implements
the first step in a transition to fully generating the plates in
OpenSCAD by adding a new plate() module. Currently, the module only
generates the basic shape of the fron (i.e., visible) portion of the
plate. This basic shape is a very close approximation of the basic
shape from the original STL files, and its width scales automatically
based on the length of the supplied serial number string. The width
was tested and confirmed to match with the width of both the 6- and
7-character STL variants. Since the width of the base plate no longer
depends upon pre-generated STL files, it will be able to continue
scaling as serial numbers reach 8 or more characters, without having
to generate new static assets.