Add rear protrusion
The original STL files include a 13mm tall, 5mm deep protrusion behind
the plate. Presumably this protrusion is to allow for clearance so that
the plate can be mounted over an enclosure panel which is mounted on the
same extrusion. Recreate the geometry of this protrusion in the fully
parameterized panel() module.

The depth of the protrusion is controlled by a new module parameter: it
can be set to 0 for mounting the plate flush to the extrusion, e.g. on
an interior facing side or unenclosed printer. It can also be adjusted
to accommodate thicker or thinner enclosure panels.

With this change, all of the geometry of the original STLs is recreated,
apart from cutouts in the rear of the plate which are not visible when
the plate is mounted. Let the slicer do its job with infill for now; if
the cutouts are really needed, it should be reasonably easy to add them.